Revert "Refactor: separate SplittableSplitAndSizeRestriction"
This reverts commit 3b09453db15a5959c4109a05e364a8a3fd4b4734.

There appears to be some inconcistency wherein the URN is registered
more than once in some downstream prod jobs.
