+++ sdks/java/harness/build.gradle
@@ -45,6 +45,7 @@ applyJavaNature(
     'AssignWindowsRunner': 'https://github.com/typetools/checker-framework/issues/3794',
     'WindowMergingFnRunner': 'https://github.com/typetools/checker-framework/issues/3794',
     'FnApiDoFnRunner': 'https://github.com/typetools/checker-framework/issues/5436',
+    'SplittablePairWithRestrictionDoFnRunner': 'https://github.com/typetools/checker-framework/issues/5436',
   ],
   automaticModuleName: 'org.apache.beam.fn.harness',
   testShadowJar: true,
