+++ sdks/java/harness/src/main/java/org/apache/beam/fn/harness/SplittablePairWithRestrictionDoFnRunner.java
@@ -58,6 +58,9 @@ import org.joda.time.Instant;
  * WindowedValue<InputT>} to {@code WindowedValue<KV<InputT, KV<RestrictionT,
  * WatermarkEstimatorStateT>>>}
  */
+@SuppressWarnings({
+  "nullness" // TODO(https://github.com/apache/beam/issues/20497)
+})
 public class SplittablePairWithRestrictionDoFnRunner<
         InputT, RestrictionT, WatermarkEstimatorStateT, OutputT>
     implements FnApiStateAccessor.MutatingStateContext<Void, BoundedWindow> {
@@ -82,17 +85,12 @@ public class SplittablePairWithRestrictionDoFnRunner<
     }
 
     private void addRunnerForPairWithRestriction(Context context) throws IOException {
-      FnApiStateAccessor<Void> stateAccessor =
-          FnApiStateAccessor.Factory.<Void>factoryForPTransformContext(context).create();
-
       SplittablePairWithRestrictionDoFnRunner<?, ?, ?, ?> runner =
           new SplittablePairWithRestrictionDoFnRunner<>(
               context.getPipelineOptions(),
               context.getPTransform(),
               context::getPCollectionConsumer,
-              stateAccessor);
-
-      stateAccessor.setKeyAndWindowContext(runner);
+              FnApiStateAccessor.Factory.factoryForPTransformContext(context));
 
       // Register processing methods
       context.addPCollectionConsumer(
@@ -123,7 +121,7 @@ public class SplittablePairWithRestrictionDoFnRunner<
       PipelineOptions pipelineOptions,
       PTransform pTransform,
       Function<String, FnDataReceiver<WindowedValue<?>>> getPCollectionConsumer,
-      FnApiStateAccessor<Void> stateAccessor)
+      FnApiStateAccessor.Factory<Void> stateAccessorFactory)
       throws IOException {
     this.pipelineOptions = pipelineOptions;
 
@@ -146,7 +144,7 @@ public class SplittablePairWithRestrictionDoFnRunner<
     FnDataReceiver<WindowedValue<KV<InputT, KV<RestrictionT, WatermarkEstimatorStateT>>>>
         mainOutputConsumer =
             (FnDataReceiver)
-                getPCollectionConsumer.apply(pTransform.getOutputsOrThrow(mainOutputTag.getId()));
+                getPCollectionConsumer.apply(pTransform.getOutputsMap().get(mainOutputTag.getId()));
     this.mainOutputConsumer = mainOutputConsumer;
 
     // Side inputs
@@ -154,16 +152,14 @@ public class SplittablePairWithRestrictionDoFnRunner<
 
     // Register processing methods
     this.observesWindow =
-        (doFnSignature.getInitialRestriction() != null
-                && doFnSignature.getInitialRestriction().observesWindow())
-            || !sideInputMapping.isEmpty();
+        doFnSignature.getInitialRestriction().observesWindow() || !sideInputMapping.isEmpty();
 
     if (observesWindow) {
       this.mutableArgumentProvider = new WindowObservingProcessBundleContext();
     } else {
       this.mutableArgumentProvider = new NonWindowObservingProcessBundleContext();
     }
-    this.stateAccessor = stateAccessor;
+    this.stateAccessor = stateAccessorFactory.create(this);
   }
 
   private void processElement(WindowedValue<InputT> elem) {
@@ -179,12 +175,15 @@ public class SplittablePairWithRestrictionDoFnRunner<
     try {
       RestrictionT currentRestriction =
           doFnInvoker.invokeGetInitialRestriction(mutableArgumentProvider);
-      WatermarkEstimatorStateT watermarkEstimatorState =
-          doFnInvoker.invokeGetInitialWatermarkEstimatorState(mutableArgumentProvider);
       outputTo(
           mainOutputConsumer,
           elem.withValue(
-              KV.of(elem.getValue(), KV.of(currentRestriction, watermarkEstimatorState))));
+              KV.of(
+                  elem.getValue(),
+                  KV.of(
+                      currentRestriction,
+                      doFnInvoker.invokeGetInitialWatermarkEstimatorState(
+                          mutableArgumentProvider)))));
     } finally {
       mutableArgumentProvider.currentElement = null;
     }
@@ -199,12 +198,15 @@ public class SplittablePairWithRestrictionDoFnRunner<
         mutableArgumentProvider.currentWindow = boundedWindow;
         RestrictionT currentRestriction =
             doFnInvoker.invokeGetInitialRestriction(mutableArgumentProvider);
-        WatermarkEstimatorStateT watermarkEstimatorState =
-            doFnInvoker.invokeGetInitialWatermarkEstimatorState(mutableArgumentProvider);
         outputTo(
             mainOutputConsumer,
             WindowedValue.of(
-                KV.of(elem.getValue(), KV.of(currentRestriction, watermarkEstimatorState)),
+                KV.of(
+                    elem.getValue(),
+                    KV.of(
+                        currentRestriction,
+                        doFnInvoker.invokeGetInitialWatermarkEstimatorState(
+                            mutableArgumentProvider))),
                 elem.getTimestamp(),
                 boundedWindow,
                 elem.getPane()));
@@ -302,11 +304,7 @@ public class SplittablePairWithRestrictionDoFnRunner<
 
     @Override
     public Object sideInput(String tagId) {
-      PCollectionView<Object> pCollectionView =
-          (PCollectionView<Object>)
-              checkStateNotNull(sideInputMapping.get(tagId), "Side input tag not found: %s", tagId);
-
-      return stateAccessor.get(pCollectionView, getCurrentWindow());
+      return stateAccessor.get(sideInputMapping.get(tagId), getCurrentWindowOrFail());
     }
   }
